==> Step 2 Data Cleaning.ipynb <==
import pandas as pd
import numpy as np 
import matplotlib.pyplot as plt
import seaborn as sns

from google.colab import drive
drive.mount('/content/drive')

df = pd.read_parquet('/content/drive/MyDrive/Elective_Groups/Capstone/Step_1_transactions_dataset_month.parquet')
df.info()

df

dash = df.groupby('client_id').count().reset_index()[['client_id','date_order']]
dash

dash.date_order.describe()

max_total_order_count = dash.quantile(0.95)
max_total_order_count

dash = dash.loc[dash['date_order'] < float(max_total_order_count)].loc[dash['date_order'] > 2]
dash

df = df.loc[df.client_id.isin(dash.client_id)].reset_index(drop=True)
df

df.to_parquet('/content/drive/MyDrive/Elective_Groups/Capstone/Step_2_transactions_dataset_outlier_removal.parquet')
