==> Step 3 Add_Segments.ipynb <==
import pandas as pd
import numpy as np 
import matplotlib.pyplot as plt
import seaborn as sns

from google.colab import drive
drive.mount('/content/drive')

df = pd.read_parquet('/content/drive/MyDrive/Elective_Groups/Capstone/Step_2_transactions_dataset_outlier_removal.parquet')
df.info

df.head()

df_customer = pd.DataFrame()

model_customer = df[['client_id','date_order']].groupby('client_id').count().reset_index()
model_customer = model_customer.loc[model_customer.date_order > 0]
df_customer['client_id'] = model_customer['client_id']
df_customer.head()

1.1 Recency

max_date = df[['client_id','date_order']].groupby('client_id').max().reset_index()
df_customer = df_customer.merge(max_date,on='client_id',how='left')
df_customer.head()

import datetime
testeddate = '2019/09/22'
NOW = datetime.datetime.strptime(testeddate,'%Y/%m/%d')

df_customer["Recency"] = df_customer["date_order"].apply(lambda x: (NOW - x).days)
df_customer

1.2 Frequency

#amount_track = df.groupby('client_id').agg({'quantity':['sum','mean'],'sales_net': ['count','sum','mean']}).reset_index()
amount_track = df[['client_id','quantity']].groupby('client_id').sum().reset_index()
amount_track = amount_track.rename(columns={'quantity':'Total_Quantity'})
df_customer = df_customer.merge(amount_track,on='client_id',how='left')
df_customer

#amount_track = df.groupby('client_id').agg({'quantity':['sum','mean'],'sales_net': ['count','sum','mean']}).reset_index()
refund_amount_track = df.loc[df.sales_net < 0][['client_id','quantity']].groupby('client_id').sum().reset_index()
refund_amount_track = refund_amount_track.rename(columns={'quantity':'R_Total_Quantity'})
df_customer = df_customer.merge(refund_amount_track,on='client_id',how='left')
df_customer

df_customer["Frequency"] = df_customer["Total_Quantity"]-df_customer["R_Total_Quantity"]
df_customer.head()

1.3 Monetary Value

#amount_track = df.groupby('client_id').agg({'quantity':['sum','mean'],'sales_net': ['count','sum','mean']}).reset_index()
sales_total = df[['client_id','sales_net']].groupby('client_id').sum().reset_index()
sales_total = sales_total.rename(columns={'sales_net':'Total_Sales'})
df_customer = df_customer.merge(sales_total,on='client_id',how='left')
df_customer

#amount_track = df.groupby('client_id').agg({'quantity':['sum','mean'],'sales_net': ['count','sum','mean']}).reset_index()
refund_sales_total = df.loc[df.sales_net < 0][['client_id','sales_net']].groupby('client_id').sum().reset_index()
refund_sales_total = refund_sales_total.rename(columns={'sales_net':'R_Total_Sales'})
df_customer = df_customer.merge(refund_sales_total,on='client_id',how='left')
df_customer

df_customer["Monetary_Value"] = df_customer["Total_Sales"]+df_customer["R_Total_Sales"]

R_2 = df_customer["Recency"].quantile(0.2)
R_4 = df_customer["Recency"].quantile(0.4)
R_6 = df_customer["Recency"].quantile(0.6)
R_8 = df_customer["Recency"].quantile(0.8)

def Recency_Quantitle(value):
  if value < R_2:
    return 5
  elif value < R_4:
    return 4
  elif value < R_6:
    return 3
  elif value < R_8:
    return 2
  else:
    return 1

F_2 = df_customer["Frequency"].quantile(0.2)
F_4 = df_customer["Frequency"].quantile(0.4)
F_6 = df_customer["Frequency"].quantile(0.6)
F_8 = df_customer["Frequency"].quantile(0.8)

def Frequency_Quantitle(value):
  if value < F_2:
    return 5
  elif value < F_4:
    return 4
  elif value < F_6:
    return 3
  elif value < F_8:
    return 2
  else:
    return 1

M_2 = df_customer["Monetary_Value"].quantile(0.2)
M_4 = df_customer["Monetary_Value"].quantile(0.4)
M_6 = df_customer["Monetary_Value"].quantile(0.6)
M_8 = df_customer["Monetary_Value"].quantile(0.8)

def Monetary_Value_Quantitle(value):
  if value < M_2:
    return 5
  elif value < M_4:
    return 4
  elif value < M_6:
    return 3
  elif value < M_8:
    return 2
  else:
    return 1

df_customer["Recency_Quantitle"] = df_customer["Recency"].apply(lambda x: Recency_Quantitle(x))
df_customer["Frequency_Quantitle"] = df_customer["Frequency"].apply(lambda x: Frequency_Quantitle(x))
df_customer["Monetary_Value_Quantitle"] = df_customer["Monetary_Value"].apply(lambda x: Monetary_Value_Quantitle(x))

df_customer['RFM'] = df_customer['Recency_Quantitle']*100 + df_customer['Frequency_Quantitle']*10 +df_customer['Monetary_Value_Quantitle']

def Score_Quantitle(value):
    if value in [555, 554, 544, 545, 454, 455, 445]:
        return "Champion"
    elif value in [543, 444, 435, 355, 354, 345, 344, 335]:
        return "Loyal Customer"
    elif value in [553, 551, 552, 541, 542, 533, 532, 531, 452, 451, 442, 441, 431, 453, 433, 432, 423, 353, 352, 351, 342, 341, 333, 323]:
        return "Potential Loyalist"
    elif value in [512, 511, 422, 421, 412, 411, 311]:
        return "New Customer"
    elif value in [525, 524, 523, 522, 521, 515, 514, 513, 425, 424, 413, 414, 415, 315, 314, 313]:
        return "Promising"
    elif value in [535, 534, 443, 434, 343, 334, 325, 324]:
        return "Need Attention"
    elif value in [155, 154, 144, 214,215,115, 114, 113]:
        return "Cannot Lose Them"
    elif value in [331, 321, 312, 221, 213]:
        return "About To Sleep"
    elif value in [255, 254, 245, 244, 253, 252, 243, 242, 235, 234, 225, 224, 153, 152, 145, 143, 142, 135, 134, 133, 125, 124]:
        return "At Risk"
    elif value in [332, 322, 231, 241, 251, 233, 232, 223, 222, 132, 123, 122, 212, 211]:
        return "Hibernating"
    elif value in [111, 112, 121, 131, 141, 151]:
        return "Lost"



df_customer["RFM_Segment"] = df_customer["RFM"].apply(lambda x: Score_Quantitle(x))

df_customer

df_customer.to_csv('/content/drive/MyDrive/Elective_Groups/Capstone/Step_3_customer_dashboard.csv')
