==> Step 4 Turn into Monthly Series.ipynb <==
import pandas as pd
import numpy as np 
import matplotlib.pyplot as plt
import seaborn as sns

from google.colab import drive
drive.mount('/content/drive')

df = pd.read_parquet('/content/drive/MyDrive/Elective_Groups/Capstone/Step_2_transactions_dataset_outlier_removal.parquet')
df.info

df_customer = pd.read_csv('/content/drive/MyDrive/Elective_Groups/Capstone/Step_3_customer_dashboard.csv')
df_customer["client_id"]=df_customer["client_id"].values.astype(str)
df_customer.head()

df.head()

unique_month = df[['month']].drop_duplicates().sort_values(by='month').month.tolist()

eval_group = pd.DataFrame()
eval_group['Target_Group'] = pd.Series(['Total Customer Traffic','Total Cutomers Distinct','Total New Cutomers Traffic','Total New Cutomers Distinct'])
eval_group['key'] = 0
eval_group.head()

cus_dash = pd.DataFrame()
unique_customer_segment = df_customer[['RFM_Segment']].drop_duplicates().RFM_Segment.tolist()
cus_dash['Segments'] = unique_customer_segment
cus_dash['key'] = 0
cus_dash.head()

dashboard = cus_dash.merge(eval_group,on='key',how='outer')[['Segments','Target_Group']]
dashboard

seg_dict = {}
for seg in unique_customer_segment:
  seg_dict[seg] = set()
seg_dict

test start

monthly_data = df.loc[df.month == '2017-09']
monthly_data

segdata = monthly_data.merge(df_customer[['client_id','RFM_Segment']],on='client_id',how='left')
for seg in unique_customer_segment:
  print(len(segdata.loc[segdata.RFM_Segment == seg]))


test end

from zmq.constants import TCP_ACCEPT_FILTER
count = 0
customer_database = set()

for month in unique_month:
  monthly_data = df.loc[df.month == month]
  print(month)
  dash_row = []
  #tct = len(monthly_data.client_id)
  #dash_row.append(tct)
  #print('Total Customer Traffic: ', tct)
  #tcd = len(monthly_data.client_id.unique())
  #dash_row.append(tcd)
  #print('Total Cutomers Distinct: ', tcd)
  #new_customer_base = monthly_data.loc[~monthly_data.client_id.isin(customer_database)]
  #tnct = len(new_customer_base.client_id)
  #dash_row.append(tnct)
  #print('Total New Cutomers Traffic: ', tnct)
  #tncd = len(new_customer_base.client_id.unique())
  #dash_row.append(tncd)
  #print('Total New Cutomers Distinct: ', tncd)
  
  segdata = monthly_data.merge(df_customer[['client_id','RFM_Segment']],on='client_id',how='left')
  for seg in unique_customer_segment:
    cus_sub = segdata.loc[segdata.RFM_Segment == seg]
    tct = len(cus_sub)
    tcd = len(cus_sub.client_id.unique())
    new_customer_base = segdata.loc[~segdata.client_id.isin(customer_database)]
    new_cus_sub = new_customer_base.loc[new_customer_base.RFM_Segment == seg]
    tnct = len(new_cus_sub)
    tncd = len(new_cus_sub.client_id.unique())
    print(seg, 'Total Count: ',tct)
    print(seg, 'Total Distinct Count: ',tcd)
    print(seg, 'Total New Count: ',tnct)
    print(seg, 'Total New Distinct Count: ',tncd)
    dash_row.append(tct)
    dash_row.append(tcd)
    dash_row.append(tnct)
    dash_row.append(tncd)
  customer_database = customer_database.union(set(monthly_data.client_id))
  dashboard[month] = pd.Series(dash_row)

dashboard

dashboard.to_csv('/content/drive/MyDrive/Elective_Groups/Capstone/Step_4_dashboard_monthly.csv',index=False)
